refactor(ui): handling swrv data fetching within higher-order mew list & mew components, split tags pages into seperate components
