+++ ui/src/components/BaseMewList.vue
@@ -1,10 +1,10 @@
 <template>
   <h6 v-if="title" class="q-mb-md">{{ title }}</h6>
 
-  <MewListSkeleton v-if="isLoading" />
-  <EmptyMewsFeed v-else-if="!isLoading && feedMews.length === 0" />
+  <MewListSkeleton v-if="isLoading || !feedMews" />
+  <EmptyMewsFeed v-else-if="feedMews && !isLoading && feedMews.length === 0" />
   <QList v-else bordered separator>
-    <MewListItem
+    <BaseMewListItem
       v-for="(mew, index) of feedMews"
       :key="index"
       :feed-mew="mew"
@@ -15,18 +15,19 @@
 
 <script setup lang="ts">
 import { QList } from "quasar";
-import MewListItem from "@/components/MewListItem.vue";
+import BaseMewListItem from "@/components/BaseMewListItem.vue";
 import MewListSkeleton from "@/components/MewListSkeleton.vue";
 import { FeedMew } from "@/types/types";
 import EmptyMewsFeed from "./EmptyMewsFeed.vue";
 
 withDefaults(
   defineProps<{
-    feedMews: FeedMew[];
+    feedMews?: FeedMew[];
     isLoading: boolean;
     title?: string;
   }>(),
   {
+    feedMews: undefined,
     title: undefined,
   }
 );
