+++ ui/src/components/CreateMewForm.vue
@@ -36,7 +36,10 @@
     </QCardSection>
 
     <QCardSection>
-      <CreateMewField :mew-type="mewType" v-bind="$attrs" />
+      <CreateMewField
+        :mew-type="mewType"
+        @publish-mew="(val) => emit('publish-mew', val)"
+      />
     </QCardSection>
   </QCard>
 </template>
@@ -53,6 +56,7 @@ defineProps<{
   originalMew?: FeedMew;
   originalAuthor?: Profile;
 }>();
+const emit = defineEmits(["publish-mew"]);
 </script>
 
 <style lang="sass" scoped>
